Add files via upload
This is the first draft. I will be adding more to it as time goes on. Because I have 2 aneurysms, I am forced to move ahead with my work much faster than usual. It is my hope that the right people get interested in this theory so that we can fix some the problems that we got ourselves into. Enjoy!